Add files via upload
These files are formatted to cut 6mm sheets and can be used with any standard laser-cutter. If you plan to use other thicknesses you will have to update the slits to fit that thickness. Note that for the connectors you will need to cut 2 bases and 4-6 connectors per piece.

Attached also an image cardboard prototype for reference.